Synchronize text view and outliner row heights

+++ OutlinerView/OutlinerRowModel.swift
@@ -7,11 +7,12 @@
 
 import Foundation
 
+/// A single row in the outline: its text, depth, expansion state, and measured height.
 struct OutlinerRowModel: Identifiable {
 	let id = UUID()
 	var text: String
 	var depth: Int
 	var isExpanded: Bool
 	var hasChildren: Bool
-	var height: Int
+	var height: CGFloat
 }
